Add department selection UI
